requirements.txt
pyFluidSynth>=1.3.4
numpy>=1.26
asyncio
InquirerPy